requirements.txt
streamlit==1.31.1
tensorflow-cpu==2.13.0
numpy
pillow
